Split optimized YUV implementations to separate files.
